refactoring

+++ src/main/java/ru/homecredit/confrestextension/web/AttachmentController.java
@@ -1,21 +1,27 @@
 package ru.homecredit.confrestextension.web;
 
+import com.atlassian.confluence.core.ContentPermissionManager;
 import com.atlassian.confluence.pages.AttachmentManager;
+import com.atlassian.confluence.security.SpacePermissionManager;
+import com.atlassian.confluence.user.UserAccessor;
 import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
-import com.atlassian.plugins.rest.common.security.AnonymousAllowed;
+import com.atlassian.sal.api.user.UserManager;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import lombok.extern.slf4j.Slf4j;
 import ru.homecredit.confrestextension.response.AttachmentResponse;
 import ru.homecredit.confrestextension.service.AttachmentService;
+import ru.homecredit.confrestextension.service.PermissionService;
 
 import javax.inject.Inject;
 import javax.inject.Named;
+import javax.persistence.criteria.CriteriaBuilder;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.util.HashMap;
-import java.util.Map;
+
+import static ru.homecredit.confrestextension.response.AttachmentResponse.Result.ERROR;
+import static ru.homecredit.confrestextension.response.AttachmentResponse.Result.SUCCESS;
 
 /**
  * A resource of message.
@@ -25,35 +31,61 @@ import java.util.Map;
 @Slf4j
 public class AttachmentController {
     private final AttachmentService attachmentService;
+    private final PermissionService permissionService;
     private final Gson gson = new GsonBuilder().setPrettyPrinting().create();
 
     @Inject
-    public AttachmentController(@ComponentImport AttachmentManager attachmentManager) {
+    public AttachmentController(@ComponentImport AttachmentManager attachmentManager,
+                                @ComponentImport UserManager userManager,
+                                @ComponentImport SpacePermissionManager spacePermissionManager,
+                                @ComponentImport UserAccessor userAccessor,
+                                @ComponentImport ContentPermissionManager contentPermissionManager) {
         log.warn("AttachmentController instance construction");
         attachmentService = new AttachmentService(attachmentManager);
+        permissionService = new PermissionService(userManager,
+                                                  spacePermissionManager,
+                                                  userAccessor,
+                                                  attachmentService,
+                                                  contentPermissionManager);
     }
 
     @GET
-    @AnonymousAllowed
     @Produces({MediaType.APPLICATION_JSON})
     @Path("/{attachmentId}/getversions")
     public Response getAttachmentVersions(@PathParam("attachmentId") String attachmentId) {
         log.info("starting getAttachmentVersions() method");
         AttachmentResponse attachmentResponse =
-                attachmentService.getVersions(Long.parseLong(attachmentId));
+        attachmentService.getVersions(Long.parseLong(attachmentId));
+        if (attachmentResponse.getResult().equals(SUCCESS) &&
+                !permissionService.hasPermission(Long.parseLong(attachmentId))) {
+            log.error("permission denied");
+            attachmentResponse.setResult(ERROR);
+            attachmentResponse.setMessage("user {} is not admin and not authorized to delete this" +
+                                     " attachment");
+            attachmentResponse.setVersions(null);
+        }
         return Response.ok(gson.toJson(attachmentResponse)).build();
     }
 
     @DELETE
-    @AnonymousAllowed
     @Produces({MediaType.APPLICATION_JSON})
     @Path("/{attachmentId}/deleteversion/{versionId}")
     public Response deleteVersion(@PathParam("attachmentId") String attachmentId,
                                   @PathParam("versionId") String versionId) {
         log.info("starting deleteVersion() method");
         AttachmentResponse attachmentResponse =
-                attachmentService.deleteVersion(Long.parseLong(attachmentId),
-                                                Integer.parseInt(versionId));
-        return Response.ok(attachmentResponse).build();
+                attachmentService.getVersions(Long.parseLong(attachmentId));
+        if (attachmentResponse.getResult() == SUCCESS) {
+            if (!permissionService.hasPermission(Long.parseLong(attachmentId))) {
+                attachmentResponse.setResult(ERROR);
+                attachmentResponse.setMessage("nor the user is an admin and is not authorized to " +
+                                                      "delete this attachment either");
+                attachmentResponse.setVersions(null);
+            } else {
+                attachmentResponse = attachmentService.deleteVersion(Long.parseLong(attachmentId),
+                                                                     Integer.parseInt(versionId));
+            }
+        }
+        return Response.ok(gson.toJson(attachmentResponse)).build();
     }
 }
